Feature: improving staging of files + unit tests

+++ test/testStageAll.sh
@@ -3,7 +3,7 @@
 
 . ./infra.inc
 
-echo "*** Running test StageAll"
+echo "*** Running test StageAll -- positive test"
 
 rm -rf stageall
 mkdir stageall && cd stageall
@@ -22,6 +22,24 @@ if [ `git status -s | grep "A  dummy.txt" | wc -l`  != 1  ]; then
 	echo "ERROR: auto-stating did not work; this is the status of the current git repo:"
 	git status
 fi
+
+cd ..
+rm -rf stageall
+
+
+echo "*** Running test StageAll -- negative test - no file available"
+
+rm -rf stageall
+mkdir stageall && cd stageall
+createSimpleRepo
+
+$SUBJECT -s
+
+if [ $? != 0 ]; then
+	echo "ERROR: Non-zero exit on test"
+	exit 1
+fi
+
 
 cd ..
 rm -rf stageall
